requirements.txt

whitenoise==6.8.2
dj-database-url==2.3.0
psycopg2-binary==2.9.10
gunicorn==23.0.0

django-redis==5.4.0

django-allauth==65.3.1

requests==2.32.3

PyJWT==2.10.1

django-tailwind==3.8.0
